backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0

django-cors-headers>=3.13,<5.0
django-environ>=0.10.0
psycopg2-binary
